move "open browser" to tools/web/webaddons.py
